New refactoring code 10.04.2023 - part 2

+++ src/main/resources/liquibase/scripts/ads_comment.sql
@@ -5,9 +5,9 @@
 CREATE TABLE IF NOT EXISTS ads_comment
 (
     id         BIGSERIAL PRIMARY KEY,
-    author_id  integer REFERENCES users (id),
+    author_id  BIGINT REFERENCES users (id),
     created_at timestamp,
     text       text,
-    ads_id     integer REFERENCES ads (id)
-
+    ads_id     BIGINT REFERENCES ads (id)
 );
+
